# remove the lock of global module for logging on RedisAI_RunSession