--- a/src/redisai.c
+++ b/src/redisai.c
@@ -730,34 +730,16 @@ void *RedisAI_RunSession(void *arg) {
   struct RedisAI_RunInfo *rinfo = (struct RedisAI_RunInfo*)arg;
   rinfo->err = RedisModule_Calloc(1, sizeof(RAI_Error));
 
-  mstime_t start, end;
   if (rinfo->mctx) {
-    start = mstime();
     rinfo->status = RAI_ModelRun(rinfo->mctx, rinfo->err);
-    end = mstime();
   }
   else if (rinfo->sctx) {
-    start = mstime();
     rinfo->status = RAI_ScriptRun(rinfo->sctx, rinfo->err);
-    end = mstime();
   }
 
-  if (rinfo->client == NULL) {
-    return NULL;
+  if (rinfo->client != NULL) {
+      RedisModule_UnblockClient(rinfo->client, rinfo);
   }
-
-  RedisModuleCtx *ctx = RedisModule_GetThreadSafeContext(rinfo->client);
-  RedisModule_ThreadSafeContextLock(ctx);
-  if (rinfo->err->code != RAI_OK) {
-    RedisModule_Log(ctx, "warning", "ERR %s", rinfo->err->detail);
-  }
-  else{
-    RedisModule_Log(ctx, "verbose", "Run took %fms", (end - start));
-  }
-  RedisModule_ThreadSafeContextUnlock(ctx);
-
-  RedisModule_UnblockClient(rinfo->client, rinfo);
-  RedisModule_FreeThreadSafeContext(ctx);
   return NULL;
 }
 
@@ -774,7 +756,8 @@ int RedisAI_Run_Reply(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
   struct RedisAI_RunInfo *rinfo = RedisModule_GetBlockedClientPrivateData(ctx);
 
   if (rinfo->status) {
-    int ret = RedisModule_ReplyWithError(ctx, rinfo->err->detail_oneline);
+      RedisModule_Log(ctx, "warning", "ERR %s", rinfo->err->detail);
+      int ret = RedisModule_ReplyWithError(ctx, rinfo->err->detail_oneline);
     RedisAI_FreeRunInfo(ctx, rinfo);
     return ret;
   }