# Changes on this PR


- [add] remove the lock of global module for logging on RedisAI_RunSession ( logging is done on the callback )

**Summary Report**: Moving the error logging from the worker thread to the main thread improved or performance in around 15/20%. Not due to the logging part but due to the removal of acquiring the module global lock on every ModelRun/ScriptRun.


-------------

# Benchmark

## On client measured Overall Inferences per Second

On client measured Overall Inferences per Second | current master (fa21096a7043e5458c47ee422dcb0e0218bb5642) | perf-runsession-lock branch |   |  
-- | -- | -- | -- | --
Test Case | RedisAI master branch + TensorFlow Library: 1.14.0 + OSS redis-server 5.0.5 ( 8 clients ) | RedisAI perf-runsession-lock branch + TensorFlow Library: 1.14.0 + OSS redis-server 5.0.5 ( 8 clients ) | Difference | % Improvement (*)
creditcard-fraud dataset from Kaggle ( 284K inferences ) | 11532 | 14296 | 2764 | 23.97%



## On client measured q50 total inference time (ms)

On client measured q50 total inference time (ms) | current master (fa21096a7043e5458c47ee422dcb0e0218bb5642) | perf-runsession-lock branch |   |  
-- | -- | -- | -- | --
Test Case | TensorFlow ModelServer: 1.14.0 + TensorFlow Library: 1.14.0 + OSS redis-server 5.0.5 ( 8 clients ) | RedisAI perf-runsession-lock branch + TensorFlow Library: 1.14.0 + OSS redis-server 5.0.5 ( 8 clients ) | Difference | % Improvement (*)
creditcard-fraud dataset from Kaggle ( 284K inferences ) | 0.67 | 0.54 | 0.13 | 19.40%





### Measurement conditions
```
Hardware:

    Hardware Overview:

      Model Name: MacBook Pro
      Model Identifier: MacBookPro15,1
      Processor Name: Intel Core i7
      Processor Speed: 2.6 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Hyper-Threading Technology: Enabled
      Memory: 16 GB
```